fix : 공유 컴포넌트 타입 분리 및 오류 수정
