+++ src/shared/ui/Button/Button.types.ts
@@ -1,0 +1,9 @@
+export interface ButtonProps {
+    children: React.ReactNode;
+    fontSize: string;
+    isActive: boolean;
+    hasBorder: boolean;
+    height: string;
+    width: string;
+    onClick: () => void;
+}
